+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- improved X509 certificates by adding Subject Key Identifier and Authority Key Identifier
 
 ### Fixed